fix: update mismatch entries id in schema migration, and fixing getEntry Query

+++ db/migration/000001_init_schema.up.sql
@@ -7,7 +7,7 @@ CREATE TABLE "accounts" (
 );
 
 CREATE TABLE "entries" (
-  "id" int PRIMARY KEY,
+  "id" bigserial PRIMARY KEY,
   "account_id" bigint NOT NULL,
   "amount" bigint NOT NULL,
   "created_at" timestamptz NOT NULL DEFAULT (now())
